+++ zcash_primitives/CHANGELOG.md
@@ -20,6 +20,7 @@ and this library adheres to Rust's notion of
   - `InputSize`
   - `InputView::serialized_size`
   - `OutputView::serialized_size`
+- `zcash_primitives::transaction::component::transparent::OutPoint::txid`
 
 ### Changed
 - MSRV is now 1.70.0.